requirements.txt
accelerate
attrs
blinker
cachetools
certifi
cffi
charset-normalizer
click
clickclick
colorama
connexion
cryptography
filelock
Flask
flask-marshmallow
Flask-SQLAlchemy
Flask-WTF
fsspec
google-ai-generativelanguage
google-api-core
google-auth
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
huggingface-hub
idna
inflection
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
MarkupSafe
marshmallow
marshmallow-sqlalchemy
mpmath
networkx
numpy
packaging
pdf2image
pdfminer.six
Pillow
proto-plus
protobuf
psutil
pyasn1
pyasn1-modules
pycparser
pypdf
PyPDF2
PyYAML
referencing
regex
requests
rpds-py
rsa
safetensors
six
SQLAlchemy
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
Werkzeug
WTForms
Flask-Cors